requirements.txt
streamlit 
langchain-openai 
PyPDF2 
langchain-community 
typing-extensions
python-dotenv
faiss-cpu
openpyxl
